requirements.txt
pdfplumber
faiss-cpu
sentence-transformers
transformers
torch
nltk